Move formula composition tests out of `test_logical_meter.py`
